extract wrappers from pages to assert
