+++ src/pages/Dashboard.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import styled from 'styled-components'
 import { Link } from 'react-router-dom'
 import { Navbar } from '../components'
